sumcheck refactor & unittest to demonstrate various num_var within single sumcheck (#862)
This PR addressed few question raised from @Jiangkm3 when deal with mpcs
poly batching tasks

### change scope
- [x] refactor `virtual_polys` out from `ceno_zkvm` to mle crate so
could be share for other crates
- [x] unittest to show sumcheck with different variables
