+++ multilinear_extensions/src/lib.rs
@@ -2,6 +2,7 @@
 pub mod mle;
 pub mod util;
 pub mod virtual_poly;
+pub mod virtual_polys;
 
 #[cfg(test)]
 mod test;
