+++ sumcheck/src/test.rs
@@ -6,12 +6,82 @@ use crate::{
 };
 use ark_std::{rand::RngCore, test_rng};
 use ff_ext::{ExtensionField, FromUniformBytes, GoldilocksExt2};
-use multilinear_extensions::{mle::DenseMultilinearExtension, virtual_poly::VirtualPolynomial};
+use multilinear_extensions::{
+    mle::DenseMultilinearExtension,
+    virtual_poly::{VPAuxInfo, VirtualPolynomial},
+    virtual_polys::VirtualPolynomials,
+};
 use p3_field::PrimeCharacteristicRing;
 use rayon::iter::{IntoParallelRefMutIterator, ParallelIterator};
 use transcript::{BasicTranscript, Transcript};
 
-// TODO add more tests related to various num_vars combination after PR #162
+#[test]
+fn test_sumcheck_with_different_degree() {
+    let nv = vec![4, 5]; // test polynomial mixed with different num_var
+    test_sumcheck_with_different_degree_helper::<GoldilocksExt2>(nv);
+}
+
+fn test_sumcheck_with_different_degree_helper<E: ExtensionField>(nv: Vec<usize>) {
+    let mut rng = test_rng();
+    let degree = 2;
+    let num_multiplicands_range = (degree, degree + 1);
+    let num_products = 1;
+    // TODO investigate error when num_threads > 1
+    let num_threads = 1;
+    let mut transcript = BasicTranscript::<E>::new(b"test");
+
+    let max_num_variables = *nv.iter().max().unwrap();
+    let poly = VirtualPolynomials::new(num_threads, max_num_variables);
+
+    let input_polys = nv
+        .iter()
+        .map(|nv| {
+            VirtualPolynomial::<E>::random(*nv, num_multiplicands_range, num_products, &mut rng)
+        })
+        .collect::<Vec<_>>();
+
+    let (poly, asserted_sum) =
+        input_polys
+            .iter()
+            .fold((poly, E::ZERO), |(mut poly, sum), (new_poly, new_sum)| {
+                poly.merge(new_poly);
+                (
+                    poly,
+                    sum + E::from_u64(
+                        1 << (max_num_variables - new_poly.aux_info.max_num_variables),
+                    ) * *new_sum,
+                )
+            });
+
+    let batch_poly = poly.get_batched_polys();
+    let (proof, _) =
+        IOPProverState::<E>::prove_batch_polys(num_threads, batch_poly.clone(), &mut transcript);
+    let mut transcript = BasicTranscript::new(b"test");
+    let subclaim = IOPVerifierState::<E>::verify(
+        asserted_sum,
+        &proof,
+        &VPAuxInfo {
+            max_degree: 2,
+            max_num_variables,
+            ..Default::default()
+        },
+        &mut transcript,
+    );
+    let r = subclaim
+        .point
+        .iter()
+        .map(|c| c.elements)
+        .collect::<Vec<_>>();
+    assert_eq!(r.len(), max_num_variables);
+    assert!(
+        input_polys
+            .iter()
+            .map(|(poly, _)| { poly.evaluate(&r[..poly.aux_info.max_num_variables]) })
+            .sum::<E>()
+            == subclaim.expected_evaluation,
+        "wrong subclaim"
+    );
+}
 
 fn test_sumcheck<E: ExtensionField>(
     nv: usize,
@@ -138,16 +208,6 @@ fn test_normal_polynomial_helper<E: ExtensionField>() {
     test_sumcheck::<E>(nv, num_multiplicands_range, num_products);
     test_sumcheck_internal::<E>(nv, num_multiplicands_range, num_products);
 }
-
-// #[test]
-// fn zero_polynomial_should_error() {
-//     let nv = 0;
-//     let num_multiplicands_range = (4, 13);
-//     let num_products = 5;
-
-//     assert!(test_sumcheck(nv, num_multiplicands_range, num_products).is_err());
-//     assert!(test_sumcheck_internal(nv, num_multiplicands_range, num_products).is_err());
-// }
 
 #[test]
 fn test_extract_sum() {
